+++ src/main/java/cz/cvut/kbss/termit/service/business/VocabularyService.java
@@ -56,7 +56,14 @@ import org.springframework.web.multipart.MultipartFile;
 
 import java.net.URI;
 import java.time.Instant;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+import java.util.Optional;
+import java.util.Set;
 
 import static cz.cvut.kbss.termit.util.Constants.VOCABULARY_REMOVAL_IGNORED_RELATIONS;
 
@@ -188,12 +195,12 @@ public class VocabularyService
         return repositoryService.getRelatedVocabularies(entity);
     }
 
-    @PreAuthorize("@vocabularyAuthorizationService.canRead(#entity)")
+    @PreAuthorize("@vocabularyAuthorizationService.canRead(#vocabulary)")
     public List<RdfsStatement> getTermRelations(Vocabulary vocabulary) {
         return repositoryService.getTermRelations(vocabulary);
     }
 
-    @PreAuthorize("@vocabularyAuthorizationService.canRead(#entity)")
+    @PreAuthorize("@vocabularyAuthorizationService.canRead(#vocabulary)")
     public List<RdfsStatement> getVocabularyRelations(Vocabulary vocabulary) {
         return repositoryService.getVocabularyRelations(vocabulary, VOCABULARY_REMOVAL_IGNORED_RELATIONS);
     }
